+++ src/app/about/page.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import React from "react";
-import LinkedInIcon from "../ui/common/LinkedInIcon";
+import LinkedInIcon from "../../components/common/icons/LinkedInIcon";
 
 const teamMembers = [
   {
